fix reseted tab bar when updating swift data

+++ GOParkin9/Views/Screens/HistoryView.swift
@@ -12,16 +12,71 @@ import SwiftData
 
 struct HistoryView: View {
     
-    @Query(filter: #Predicate<ParkingRecord>{p in p.isHistory == true}) var parkingRecords: [ParkingRecord]
-    
+    @Query(
+        filter: #Predicate<ParkingRecord> { p in p.isHistory == true },
+        sort: [SortDescriptor(\.createdAt, order: .reverse)]
+    ) var allParkingRecords: [ParkingRecord]
+
+    var unpinnedParkingRecords: [ParkingRecord] {
+        allParkingRecords.filter { !$0.isPinned }
+    }
+
+    var pinnedParkingRecords: [ParkingRecord] {
+        allParkingRecords.filter { $0.isPinned }
+    }
+
     @Environment(\.modelContext) var context
 
     
     var body: some View {
         NavigationView {
             List {
-                ForEach(parkingRecords) { entry in
-                    HistoryComponent(entry: entry, pinItem: { pinItem(entry) }, deleteItem: { deleteItem(entry) })
+                
+                HStack {
+                    Image(systemName: "pin")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 20, height: 20)
+                        .padding(.trailing, 10)
+                    
+                    
+                    Text("Pinned")
+                        .font(.title3)
+                        .fontWeight(.bold)
+                        .opacity(0.6)
+                }
+                
+                if pinnedParkingRecords.isEmpty {
+                    Text("No pinned item")
+                        .foregroundColor(.secondary)
+                } else {
+                    ForEach(pinnedParkingRecords) { entry in
+                        HistoryComponent(entry: entry, pinItem: { pinItem(entry) }, deleteItem: { deleteItem(entry) })
+                    }
+                }
+                
+                HStack {
+                    Image(systemName: "clock.arrow.circlepath")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 20, height: 20)
+                        .padding(.trailing, 10)
+                    
+                    
+                    Text("All History")
+                        .font(.title3)
+                        .fontWeight(.bold)
+                        .opacity(0.6)
+                }
+                
+                if unpinnedParkingRecords.isEmpty {
+                    Text("No history item")
+                        .foregroundColor(.secondary)
+                } else {
+                    
+                    ForEach(unpinnedParkingRecords) { entry in
+                        HistoryComponent(entry: entry, pinItem: { pinItem(entry) }, deleteItem: { deleteItem(entry) })
+                    }
                 }
             }
             .navigationTitle("History")
@@ -53,13 +108,18 @@ struct HistoryView: View {
     }
 
     private func pinItem(_ entry: ParkingRecord) {
-        entry.isPinned.toggle()
-        try? context.save()
+        withAnimation {
+            
+            entry.isPinned.toggle()
+            try? context.save()
+        }
     }
     
     private func deleteItem(_ entry: ParkingRecord) {
-        context.delete(entry)
-        try? context.save()
+        withAnimation {
+            context.delete(entry)
+            try? context.save()
+        }
     }
 }
 
@@ -117,7 +177,11 @@ struct HistoryComponent: View {
             Button {
                 pinItem()
             } label: {
-                Label("Pin", systemImage: "pin")
+                if entry.isPinned {
+                    Label("Unpin", systemImage: "pin.slash")
+                } else {
+                    Label("Pin", systemImage: "pin")
+                }
             }
             .tint(.yellow)
